cdk/requirements.txt
aws-cdk-lib==2.84.0
python-dotenv==0.21.0
